@@ -21,8 +21,8 @@ export class LinhasLotacaoListComponent implements OnInit {
   public dataSource: MatTableDataSource<LotacaoModel>;
   public tableSizeOptions: number[];
 
-  @ViewChild(MatPaginator) paginator: MatPaginator;
-  @ViewChild(MatSort) sort: MatSort;
+  @ViewChild(MatPaginator, { static: true }) paginator: MatPaginator;
+  @ViewChild(MatSort, { static: true }) sort: MatSort;
 
   constructor(
     private linhasLotacaoService: LinhasLotacaoService,
