@@ -21,8 +21,8 @@ export class LinhasOnibusListComponent implements OnInit {
   public dataSource: MatTableDataSource<OnibusModel>;
   public tableSizeOptions: number[];
 
-  @ViewChild(MatPaginator) paginator: MatPaginator;
-  @ViewChild(MatSort) sort: MatSort;
+  @ViewChild(MatPaginator, { static: true }) paginator: MatPaginator;
+  @ViewChild(MatSort, { static: true }) sort: MatSort;
 
   constructor(
     private linhasOnibusService: LinhasOnibusService,
